requirements.txt
absl-py==0.13.0
attrs==21.2.0
Automat==20.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
cloudpickle==1.3.0
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cycler==0.10.0
dm-tree==0.1.6
evdev==1.4.0
filelock==3.0.12
flake8==3.9.2
future==0.18.2
gltf==0.7.13
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.37.0
gym==0.20.0
hyperlink==21.0.0
idna==3.2
ijson==3.1.4
imageio==2.9.0
importlib-metadata==4.8.1
incremental==21.3.0
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.2
lz4==3.1.3
Markdown==3.3.4
matplotlib==3.4.3
mccabe==0.6.1
mkl-fft==1.3.0
mkl-service==2.4.0
msgpack==1.0.2
networkx==2.6.3
numpy==1.21.2
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.3.56
panda3d==1.10.10
panda3d-gltf==0.13
panda3d-simplepbr==0.9
pandas==1.3.3
Pillow==8.3.2
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.0.6
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
pynput==1.7.3
pyparsing==2.4.7
pyquaternion==0.9.9
python-dateutil==2.8.2
python-xlib==0.31
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
ray==1.6.0
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.9.0
rsa==4.7.2
Rtree==0.9.7
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
sh==1.14.2
Shapely==1.7.1
six
sklearn==0.0
tableprint==0.9.1
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
threadpoolctl==2.2.0
tifffile==2021.8.30
torch==1.9.0
torchfile==0.1.0
torchvision==0.10.0
tornado==6.1
trimesh==3.9.29
Twisted==21.7.0
typing-extensions==3.10.0.2
urllib3==1.26.6
visdom==0.1.8.9
wcwidth==0.2.5
websocket-client==1.2.1
Werkzeug==2.0.1
yattag==1.14.0
zipp==3.5.0
zope.interface==5.4.0
